glove_ctrled_rohand/requirements.txt
bleak==0.22.2
bleak-winrt==1.2.0
numpy==1.26.4
pymodbus==3.7.2
pyserial==3.5
keyboard==0.13.5
cvzone==1.6.1
